last updates being made in bulk. Worked on closing sessions, added sessions to the employee side ensuring it was secure enough to go on the web ofcourse excluding the ecommerce part..
altered mechanics so that only bails that are deemed sellable in the employee section can be edited. the rest cannot. so one can add bails and they wont automatically show in the ecommerce side unless deemed to be available.

updated the map to show the actual location of the store